requirements.txt
asgiref                  
Django>=3.0,<4.0
djangorestframework      
djangorestframework-simplejwt
gunicorn
pip                      
rest_framework_simplejwt
setuptools               
sqlparse                 
typing_extensions        
psycopg2-binary
django-cors-headers
whisper
django-storages
boto3
